Refactoring perf_client
